Added types in separate files
