+++ my-app/src/components/Modal/types.ts
@@ -1,0 +1,7 @@
+export type modalType = {
+  setModalOpen: Function;
+  setImgUrl: Function;
+  setPhotoTitle: Function;
+  children: React.ReactNode;
+  modalOpen: boolean;
+};
