+++ my-app/src/pages/Albums/types.ts
@@ -1,0 +1,7 @@
+export type recordType = {
+  albumId: number;
+  id: number;
+  imgUrl: string;
+  name: string;
+  title: string;
+};
