+++ my-app/src/pages/Photos/types.ts
@@ -1,0 +1,17 @@
+export type recordType = {
+  albumId: number;
+  id: number;
+  url: string;
+  name: string;
+  title: string;
+};
+
+export interface IrecordType {
+  record: {
+    albumId: number;
+    id: number;
+    imgUrl: string;
+    name: string;
+    title: string;
+  };
+}
